% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota_preyprop.R
\docType{data}
\name{biota_preyprop}
\alias{biota_preyprop}
\title{biota_preyprop data input}
\format{An object of class \code{data.frame} with 27 rows and 27 columns.}
\usage{
biota_preyprop
}
\description{
biota_preyprop data input
}
\details{
used in \code{\link{bioaccum_batch}}
}
\keyword{datasets}
