% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cntcalc.R
\name{cntcalc}
\alias{cntcalc}
\title{Format calculated contaminant concentrations}
\usage{
cntcalc(contam, constants)
}
\arguments{
\item{contam}{input contaminants from formatted user inputs}

\item{constants}{input constants from formatted user inputs}
}
\description{
Format calculated contaminant concentrations
}
