% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\title{constants data input}
\format{An object of class \code{data.frame} with 22 rows and 3 columns.}
\usage{
constants
}
\description{
constants data input
}
\keyword{datasets}
