% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contam.R
\docType{data}
\name{contam}
\alias{contam}
\title{contam data input}
\format{An object of class \code{data.frame} with 75 rows and 12 columns.}
\usage{
contam
}
\description{
contam data input
}
\keyword{datasets}
