% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalsiteassess.R
\docType{data}
\name{finalsiteassess}
\alias{finalsiteassess}
\title{finalsiteassess data input}
\format{An object of class \code{data.frame} with 20 rows and 2 columns.}
\usage{
finalsiteassess
}
\description{
finalsiteassess data input
}
\keyword{datasets}
