% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{FoodWeb_SQO}
\alias{FoodWeb_SQO}
\title{Core food web model}
\usage{
FoodWeb_SQO(NumSim, csed, cwater, cpw, log_KowTS, logkow_tempcor, EdA, EdB,
  xdoc, ddoc, xpoc, dpoc, alphapoc, alphadoc, ocsed, ds, taxa, A, B, T,
  lipid, nloc, nlom, wc, beta, betap, mo, mp, phi, kM, Wb, Cox, vss, scav,
  preyprop, cbiota, vld, vcd, vnd, vwd, GR, assimEff_1, assimEff_2,
  assimEff_3)
}
\arguments{
\item{NumSim}{}

\item{csed}{}

\item{cwater}{}

\item{cpw}{}

\item{log_KowTS}{}

\item{logkow_tempcor}{}

\item{EdA}{}

\item{EdB}{}

\item{xdoc}{}

\item{ddoc}{}

\item{xpoc}{}

\item{dpoc}{}

\item{alphapoc}{}

\item{alphadoc}{}

\item{ocsed}{}

\item{ds}{}

\item{taxa}{}

\item{A}{}

\item{B}{}

\item{lipid}{}

\item{nloc}{}

\item{nlom}{}

\item{wc}{}

\item{beta}{}

\item{betap}{}

\item{mo}{}

\item{mp}{}

\item{phi}{}

\item{kM}{}

\item{Wb}{}

\item{Cox}{}

\item{vss}{}

\item{scav}{}

\item{preyprop}{}

\item{cbiota}{}

\item{vld}{}

\item{vcd}{}

\item{vnd}{}

\item{vwd}{}

\item{GR}{}

\item{assimEff_1}{}

\item{assimEff_2}{}

\item{assimEff_3}{}
}
\description{
Core food web model
}
