% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genlognorm_fun.R
\name{genlognorm_fun}
\alias{genlognorm_fun}
\title{Generate log normal variables from mean and sd inputs}
\usage{
genlognorm_fun(nsim, X, SD)
}
\arguments{
\item{nsim}{number of simulations}

\item{X}{mean value from user input for single contaminant}

\item{SD}{SD value from user input for single contaminant}
}
\description{
Generate log normal variables from mean and sd inputs
}
\details{
http://yasai.rutgers.edu/yasai-guide-27.html, but see macros code for YASAI_sccwrp
}
