% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_sum_fun.R
\name{indic_sum_fun}
\alias{indic_sum_fun}
\title{Summary function for calculating bsaf and cbiota totals for each indicator species}
\usage{
indic_sum_fun(cbiota, contamcalc)
}
\arguments{
\item{cbiota}{cbiota output from \code{\link{bioaccum_batch}}}

\item{contamcalc}{contaminants fom inputs}
}
\description{
Summary function for calculating bsaf and cbiota totals for each indicator species
}
