% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcs_fun.R
\name{mcs_fun}
\alias{mcs_fun}
\title{MCS function}
\usage{
mcs_fun(nsim, indic_sum, mcsparms, constants)
}
\arguments{
\item{nsim}{number of MC sims}

\item{indic_sum}{output from indic_sum_fun}

\item{mcsparms}{MCS parameter inputs}

\item{constants}{constants inputs}
}
\description{
MCS function
}
