% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcs_plo.R
\name{mcs_plo}
\alias{mcs_plo}
\title{Create MCS summary plot}
\usage{
mcs_plo(mcsres, alpha = 0.4, xmax = NULL)
}
\arguments{
\item{mcsres}{output from \code{\link{mcs_fun}}}

\item{alpha}{numeric value for transparency of background fill}

\item{xmax}{numeric value for upper limit of x-axis, defaults to 95th percentile of all linkage values}
}
\description{
Create MCS summary plot
}
