% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcs_sum_fun.R
\name{mcs_sum_fun}
\alias{mcs_sum_fun}
\title{Summarize MCS results, compare with observed}
\usage{
mcs_sum_fun(mcsres)
}
\arguments{
\item{mcsres}{MC results, output from \code{\link{mcs_fun}}}
}
\description{
Summarize MCS results, compare with observed
}
