% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcsparms.R
\docType{data}
\name{mcsparms}
\alias{mcsparms}
\title{mcsparms data input}
\format{An object of class \code{data.frame} with 101 rows and 2 columns.}
\usage{
mcsparms
}
\description{
mcsparms data input
}
\details{
Data are in long format, prefix of \code{HR} is home range, \code{sed} is sediment, \code{indic} is guild species, suffix of \code{X} is mean, \code{SD} is standard deviation
}
\keyword{datasets}
