% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsedcon_mcs_fun.R
\name{modsedcon_mcs_fun}
\alias{modsedcon_mcs_fun}
\title{mcs function for sediment concentration}
\usage{
modsedcon_mcs_fun(nsim, sedmeanse, propseaf, SUF, CVBAF, indic_sum)
}
\arguments{
\item{nsim}{number of simulations}

\item{sedmeanse}{sediment mean and se values from user input for each contaminant class}

\item{propseaf}{proportion of seafood diet, output from formpropseaf}

\item{SUF}{site use factor}

\item{CVBAF}{bioaccumulation factor sd/mean from mcsparms}

\item{indic_sum}{indicator guild concentrations sums across contaminants}
}
\description{
mcs function for sediment concentration
}
