% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modtiscon_mcs_fun.R
\name{modtiscon_mcs_fun}
\alias{modtiscon_mcs_fun}
\title{mcs function for modelled tissue concentration}
\usage{
modtiscon_mcs_fun(nsim, meanse, propseaf)
}
\arguments{
\item{nsim}{number of simulations}

\item{meanse}{mean and se values from user input for each guild species and contaminant class}

\item{propseaf}{proportion of seafood diet}
}
\description{
mcs function for modelled tissue concentration
}
