% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plo_bsaf.R
\name{plo_bsaf}
\alias{plo_bsaf}
\title{plot bsaf and cbiota for contaminant}
\usage{
plo_bsaf(bsaf, cbiota, cntbsaf)
}
\arguments{
\item{bsaf}{bsaf output from \code{\link{bioaccum_batch}}}

\item{cbiota}{cbiota output from \code{\link{bioaccum_batch}}}

\item{cntbsaf}{selected contaminant}
}
\description{
plot bsaf and cbiota for contaminant
}
