% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescmb.R
\name{rescmb}
\alias{rescmb}
\title{Combine bsaf and cbiota for plot/table of contaminant}
\usage{
rescmb(bsaf, cbiota, cntbsaf = NULL)
}
\arguments{
\item{bsaf}{bsaf output from \code{\link{bioaccum_batch}}}

\item{cbiota}{cbiota output from \code{\link{bioaccum_batch}}}

\item{cntbsaf}{selected contaminant, optional}
}
\description{
Combine bsaf and cbiota for plot/table of contaminant
}
