% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqo_sum_fun.R
\name{sqo_sum_fun}
\alias{sqo_sum_fun}
\title{SQO assessment summary}
\usage{
sqo_sum_fun(wgtavg, mcsres, tischmthr, constants, finalsiteassess)
}
\arguments{
\item{wgtavg}{weighted average observed tissue concentrations from input, by contaminant category, output from \code{wgt_avg_fun}}

\item{mcsres}{output from \code{mcs_fun}}

\item{tischmthr}{lookup table for tissue chemistry thresholds}

\item{constants}{constants from user inputs and lookup table, only SCT is used (sediment linkage threshold)}

\item{finalsiteassess}{final site assessment lookup table}
}
\description{
SQO assessment summary
}
