% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suf_mcs_fun.R
\name{suf_mcs_fun}
\alias{suf_mcs_fun}
\title{site use mcs function}
\usage{
suf_mcs_fun(nsim, constants, mcsparms)
}
\arguments{
\item{nsim}{number of simulations}

\item{constants}{constants inputs}

\item{mcsparms}{MCS parameter inputs}
}
\description{
site use mcs function
}
