% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_bsaf.R
\name{tab_bsaf}
\alias{tab_bsaf}
\title{Table of bsaf and cbiota for contaminant}
\usage{
tab_bsaf(bsaf, cbiota, cntbsaf)
}
\arguments{
\item{bsaf}{bsaf output from \code{\link{bioaccum_batch}}}

\item{cbiota}{cbiota output from \code{\link{bioaccum_batch}}}

\item{cntbsaf}{selected contaminant}
}
\description{
Table of bsaf and cbiota for contaminant
}
