% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tischmthr.R
\docType{data}
\name{tischmthr}
\alias{tischmthr}
\title{tischmthr data input}
\format{An object of class \code{data.frame} with 16 rows and 3 columns.}
\usage{
tischmthr
}
\description{
tischmthr data input
}
\keyword{datasets}
