% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioaccum_batch.R
\name{bioaccum_batch}
\alias{bioaccum_batch}
\title{Run bioaccumulation model function FoodWeb_SQO in batch for all species, contaminants}
\usage{
bioaccum_batch(biota, contam, biota_preyprop, constants)
}
\arguments{
\item{biota}{input biological data}

\item{contam}{input contaminant data}

\item{biota_preyprop}{input prey proportions}

\item{constants}{input constants}
}
\description{
Run bioaccumulation model function FoodWeb_SQO in batch for all species, contaminants
}
