% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoodWeb_SQO.R
\name{FoodWeb_SQO}
\alias{FoodWeb_SQO}
\title{Core food web model}
\usage{
FoodWeb_SQO(NumSim, csed, cwater, cpw, log_KowTS, logkow_tempcor, EdA, EdB,
  xdoc, ddoc, xpoc, dpoc, alphapoc, alphadoc, ocsed, ds, taxa, A, B, T,
  lipid, nloc, nlom, wc, beta, betap, mo, mp, phi, kM, Wb, Cox, vss, scav,
  preyprop, cbiota, vld, vcd, vnd, vwd, GR, assimEff_1, assimEff_2,
  assimEff_3)
}
\arguments{
\item{NumSim}{required input}

\item{csed}{required input}

\item{cwater}{required input}

\item{cpw}{required input}

\item{log_KowTS}{required input}

\item{logkow_tempcor}{required input}

\item{EdA}{required input}

\item{EdB}{required input}

\item{xdoc}{required input}

\item{ddoc}{required input}

\item{xpoc}{required input}

\item{dpoc}{required input}

\item{alphapoc}{required input}

\item{alphadoc}{required input}

\item{ocsed}{required input}

\item{ds}{required input}

\item{taxa}{required input}

\item{A}{required input}

\item{B}{required input}

\item{T}{required input}

\item{lipid}{required input}

\item{nloc}{required input}

\item{nlom}{required input}

\item{wc}{required input}

\item{beta}{required input}

\item{betap}{required input}

\item{mo}{required input}

\item{mp}{required input}

\item{phi}{required input}

\item{kM}{required input}

\item{Wb}{required input}

\item{Cox}{required input}

\item{vss}{required input}

\item{scav}{required input}

\item{preyprop}{required input}

\item{cbiota}{required input}

\item{vld}{required input}

\item{vcd}{required input}

\item{vnd}{required input}

\item{vwd}{required input}

\item{GR}{required input}

\item{assimEff_1}{required input}

\item{assimEff_2}{required input}

\item{assimEff_3}{required input}
}
\description{
Core food web model
}
