% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{indic_sum_fun}
\alias{indic_sum_fun}
\title{Summary function for calculating bsaf and cbiota totals for each indicator species}
\usage{
indic_sum_fun(cbiota, contamcalc)
}
\arguments{
\item{cbiota}{}

\item{contamcalc}{contaminants fom inputs}

\item{bsaf}{}
}
\description{
Summary function for calculating bsaf and cbiota totals for each indicator species
}
