% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcs_fun.R
\name{mcs_fun}
\alias{mcs_fun}
\title{MCS function}
\usage{
mcs_fun(nsim, indic_sum, mcsparms, constants, propseaf = c(0, 0.5, 0, 0,
  0.5, 0, 0, 0, 0))
}
\arguments{
\item{nsim}{number of MC sims}

\item{indic_sum}{output from indic_sum_fun}

\item{mcsparms}{MCS parameter inputs}

\item{constants}{constants inputs}

\item{propseaf}{numeric vector indicating proportion of human diet for each guild species, must sum to 1 and have length 9}
}
\description{
MCS function
}
